apps/quiz-pack-api/pyproject.toml
[project]
name = "question-generator"
version = "0.1.0"
description = "Admin tool for generating and curating quiz questions"
requires-python = ">=3.11"
dependencies = [
    "quiz-shared",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "langchain>=0.3.9",
    "langchain-openai>=0.2.10",
    "gradio>=4.0.0",
    "anthropic>=0.42.0",
    "geopandas>=0.14.0",
    "matplotlib>=3.8.0",
    "cartopy>=0.22.0",
    "boto3>=1.34.0",
    "tavily-python>=0.5.0",
    "google-generativeai>=0.8.0",
]

[tool.uv.sources]
quiz-shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = [".", "../../packages/shared"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
